Transport apiKeys and providerSettings via cookies.
While doing this, introduce a simple helper function for cookies
