+++ app/components/chat/BaseChat.tsx
@@ -3,7 +3,7 @@
  * Preventing TS checks with files presented in the video for a better presentation.
  */
 import type { Message } from 'ai';
-import React, { type RefCallback, useCallback, useEffect, useState } from 'react';
+import React, { type RefCallback, useEffect, useState } from 'react';
 import { ClientOnly } from 'remix-utils/client-only';
 import { Menu } from '~/components/sidebar/Menu.client';
 import { IconButton } from '~/components/ui/IconButton';
@@ -25,7 +25,7 @@ import GitCloneButton from './GitCloneButton';
 import FilePreview from './FilePreview';
 import { ModelSelector } from '~/components/chat/ModelSelector';
 import { SpeechRecognitionButton } from '~/components/chat/SpeechRecognition';
-import type { IProviderSetting, ProviderInfo } from '~/types/model';
+import type { ProviderInfo } from '~/types/model';
 import { ScreenshotStateManager } from './ScreenshotStateManager';
 import { toast } from 'react-toastify';
 import StarterTemplates from './StarterTemplates';
@@ -109,29 +109,6 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
     const [transcript, setTranscript] = useState('');
     const [isModelLoading, setIsModelLoading] = useState<string | undefined>('all');
 
-    const getProviderSettings = useCallback(() => {
-      let providerSettings: Record<string, IProviderSetting> | undefined = undefined;
-
-      try {
-        const savedProviderSettings = Cookies.get('providers');
-
-        if (savedProviderSettings) {
-          const parsedProviderSettings = JSON.parse(savedProviderSettings);
-
-          if (typeof parsedProviderSettings === 'object' && parsedProviderSettings !== null) {
-            providerSettings = parsedProviderSettings;
-          }
-        }
-      } catch (error) {
-        console.error('Error loading Provider Settings from cookies:', error);
-
-        // Clear invalid cookie data
-        Cookies.remove('providers');
-      }
-
-      return providerSettings;
-    }, []);
-
     useEffect(() => {
       console.log(transcript);
     }, [transcript]);
@@ -171,7 +148,6 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
     useEffect(() => {
       if (typeof window !== 'undefined') {
         let parsedApiKeys: Record<string, string> | undefined = {};
-        const providerSettings = getProviderSettings();
 
         try {
           parsedApiKeys = getApiKeysFromCookies();
@@ -182,12 +158,7 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
         }
 
         setIsModelLoading('all');
-        fetch('/api/models', {
-          headers: {
-            'x-client-api-keys': JSON.stringify(parsedApiKeys),
-            'x-client-provider-settings': JSON.stringify(providerSettings),
-          },
-        })
+        fetch('/api/models')
           .then((response) => response.json())
           .then((data) => {
             const typedData = data as { modelList: ModelInfo[] };
@@ -209,26 +180,21 @@ export const BaseChat = React.forwardRef<HTMLDivElement, BaseChatProps>(
 
       setIsModelLoading(providerName);
 
+      let providerModels: ModelInfo[] = [];
+
       try {
-        const providerSettings = getProviderSettings();
-
-        const response = await fetch(`/api/models/${encodeURIComponent(providerName)}`, {
-          headers: {
-            'x-client-api-keys': JSON.stringify(newApiKeys),
-            'x-client-provider-settings': JSON.stringify(providerSettings),
-          },
-        });
+        const response = await fetch(`/api/models/${encodeURIComponent(providerName)}`);
         const data = await response.json();
-        const typedData = data as { modelList: ModelInfo[] };
-
-        // Only update models for the specific provider
-        setModelList((prevModels) => {
-          const otherModels = prevModels.filter((model) => model.provider !== providerName);
-          return [...otherModels, ...typedData.modelList];
-        });
+        providerModels = (data as { modelList: ModelInfo[] }).modelList;
       } catch (error) {
-        console.error('Error loading dynamic models:', error);
+        console.error('Error loading dynamic models for:', providerName, error);
       }
+
+      // Only update models for the specific provider
+      setModelList((prevModels) => {
+        const otherModels = prevModels.filter((model) => model.provider !== providerName);
+        return [...otherModels, ...providerModels];
+      });
       setIsModelLoading(undefined);
     };
